Fix the UI/UX
