+++ packages/chess-app/src/components/NewGame.tsx
@@ -103,7 +103,8 @@ function NewGameModalContent({
         <div className="flex flex-col items-start border rounded-lg shadow-md bg-gray-50 dark:bg-gray-700 border-gray-300 dark:border-gray-700">
           <div className="relative group w-full p-6 border-b border-gray-200 dark:border-gray-600">
             <p className="block mb-2 text-sm font-medium text-gray-900 dark:text-gray-200">
-              Share this link with your opponent to start the game.
+              Share this link with the black player. The black player will get back to you with the
+              final game link.
             </p>
             <p
               className="text-sm text-blue-600 underline cursor-pointer truncate hover:text-blue-800 dark:text-blue-400 dark:hover:text-blue-600 focus:ring-0"
